cambios en walmart2 y store

+++ Proyectos/Proyecto1/src/clients/Coupon.java
@@ -4,16 +4,16 @@ import java.util.List;
 
 public class Coupon {
 
-    private double porcentage;
+    private int porcentage;
 
     private List<Country> countries;
 
-    public Coupon(double porcentage, List<Country> countries) {
+    public Coupon(int porcentage, List<Country> countries) {
         this.porcentage = porcentage;
         this.countries = countries;
     }
 
-    public double getPorcentage() {
+    public int getPorcentage() {
         return porcentage;
     }
